+++ react-components/src/query/useModelsForInstanceQuery.ts
@@ -8,13 +8,7 @@ import { useFdmSdk, useSDK } from '../components/RevealCanvas/SDKProvider';
 import { type InternalId, type CogniteClient } from '@cognite/sdk';
 import { type TaggedAddResourceOptions } from '../components/Reveal3DResources/types';
 import { getImage360CollectionsForAsset } from '../hooks/network/getImage360CollectionsForAsset';
-import {
-  type AssetInstanceReference,
-  type InstanceReference,
-  isHybridAssetCoreDmsInstance,
-  isAssetInstance,
-  isDmsInstance
-} from '../data-providers/types';
+
 import { uniqBy } from 'lodash';
 import { createAddOptionsKey } from '../utilities/createAddOptionsKey';
 import { type Fdm3dDataProvider } from '../data-providers/Fdm3dDataProvider';
@@ -22,7 +16,12 @@ import { type DmsUniqueIdentifier } from '../data-providers';
 import { getPointCloudModelsForAssetInstance } from '../hooks/network/getPointCloudModelsForAssetInstance';
 import { type FdmSDK } from '../data-providers/FdmSDK';
 import { useFdm3dDataProvider } from '../components/CacheProvider/CacheProvider';
-import { type InstanceReference, isDmsInstance, isInternalId } from '../utilities/instanceIds';
+import {
+  type InstanceReference,
+  isDmsInstance,
+  isHybridAssetCoreDmsInstance,
+  isInternalId
+} from '../utilities/instanceIds';
 
 export const useModelsForInstanceQuery = (
   instance: InstanceReference | undefined
