+++ react-components/src/utilities/instanceIds/typeGuards.ts
@@ -3,10 +3,12 @@
  */
 import { type ExternalId, type IdEither, type InternalId } from '@cognite/sdk';
 import { type DmsUniqueIdentifier } from '../../data-providers';
-import { type InstanceReference } from './types';
+import { type AssetHybridInstanceReference, type InstanceReference } from './types';
 
 export function isIdEither(instance: InstanceReference): instance is IdEither {
-  return isExternalId(instance) || isInternalId(instance);
+  return (
+    (instance as ExternalId).externalId !== undefined || (instance as InternalId).id !== undefined
+  );
 }
 
 export function isExternalId(idEither: IdEither): idEither is ExternalId {
@@ -19,4 +21,10 @@ export function isInternalId(idEither: IdEither): idEither is InternalId {
 
 export function isDmsInstance(instance: InstanceReference): instance is DmsUniqueIdentifier {
   return 'externalId' in instance && 'space' in instance;
+}
+
+export function isHybridAssetCoreDmsInstance(
+  instance: InstanceReference
+): instance is AssetHybridInstanceReference {
+  return 'assetInstanceId' in instance && isDmsInstance(instance.assetInstanceId);
 }
